refactor(markdown): replace MCPCallElement with ToolCallElement and remove unused components
- Replaced MCPCallElement with ToolCallElement in the markdown plugins to enhance functionality.
- Removed the MCPCall component and its associated render logic to streamline the codebase.
- Added new ToolCall component to handle tool usage and results with improved structure and performance.
- Updated relevant imports and ensured adherence to performance optimization practices.
