+++ packages/ai-workspace-common/src/components/markdown/plugins/tool-call/index.ts
@@ -2,17 +2,17 @@ import Component from './render';
 import rehypePlugin, { TOOL_USE_TAG_RENDER } from './rehypePlugin';
 
 // Define a consistent interface for the plugin
-type MCPCallElement = {
+type ToolCallElement = {
   Component: typeof Component;
   rehypePlugin: typeof rehypePlugin;
   tag: string;
 };
 
 // Create the plugin object
-const MCPCallElement: MCPCallElement = {
+const ToolCallElement: ToolCallElement = {
   Component,
   rehypePlugin,
   tag: TOOL_USE_TAG_RENDER, // We'll use TOOL_USE_TAG_RENDER as our main tag since we'll handle both tool_use and tool_use_result
 };
 
-export default MCPCallElement;
+export default ToolCallElement;
